Extract interfaces for ClusteredTierClientEntity
