+++ clustered/client/src/main/java/org/ehcache/clustered/client/internal/store/InternalClusterTierClientEntity.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright Terracotta, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.ehcache.clustered.client.internal.store;
+
+import org.ehcache.clustered.client.internal.Timeouts;
+
+import java.util.UUID;
+
+/**
+ * InternalClusterTierClientEntity
+ */
+public interface InternalClusterTierClientEntity extends ClusteredTierClientEntity {
+
+  void setTimeouts(Timeouts timeouts);
+
+  void setClientId(UUID clientId);
+
+  void setStoreIdentifier(String storeIdentifier);
+
+}
