+++ clustered/client/src/test/java/org/ehcache/clustered/client/internal/service/ClusteredStateRepositoryReplicationTest.java
@@ -17,25 +17,20 @@
 package org.ehcache.clustered.client.internal.service;
 
 import org.ehcache.clustered.client.config.ClusteringServiceConfiguration;
-import org.ehcache.clustered.client.config.builders.ClusteredResourcePoolBuilder;
 import org.ehcache.clustered.client.config.builders.ClusteringServiceConfigurationBuilder;
-import org.ehcache.clustered.client.internal.EhcacheClientEntity;
 import org.ehcache.clustered.client.internal.EhcacheClientEntityService;
 import org.ehcache.clustered.client.internal.UnitTestConnectionService;
 import org.ehcache.clustered.client.internal.lock.VoltronReadWriteLockEntityClientService;
-import org.ehcache.clustered.client.internal.store.ClusteredTierClientEntity;
 import org.ehcache.clustered.client.internal.store.ClusteredTierClientEntityService;
 import org.ehcache.clustered.client.internal.store.ServerStoreProxy;
+import org.ehcache.clustered.client.internal.store.SimpleClusteredTierClientEntity;
 import org.ehcache.clustered.client.service.ClusteringService;
 import org.ehcache.clustered.common.Consistency;
 import org.ehcache.clustered.lock.server.VoltronReadWriteLockServerEntityService;
 import org.ehcache.clustered.server.EhcacheServerEntityService;
 import org.ehcache.clustered.server.store.ClusteredTierServerEntityService;
-import org.ehcache.config.builders.ResourcePoolsBuilder;
 import org.ehcache.core.config.BaseCacheConfiguration;
-import org.ehcache.core.config.DefaultConfiguration;
 import org.ehcache.core.internal.store.StoreConfigurationImpl;
-import org.ehcache.spi.persistence.PersistableResourceService;
 import org.ehcache.spi.persistence.StateHolder;
 import org.junit.After;
 import org.junit.Before;
@@ -43,7 +38,6 @@ import org.junit.Test;
 import org.terracotta.offheapresource.OffHeapResourcesProvider;
 import org.terracotta.offheapresource.config.MemoryUnit;
 import org.terracotta.passthrough.PassthroughClusterControl;
-import org.terracotta.passthrough.PassthroughServer;
 import org.terracotta.passthrough.PassthroughTestHelpers;
 
 import java.lang.reflect.Field;
@@ -107,7 +101,7 @@ public class ClusteredStateRepositoryReplicationTest {
 
     ServerStoreProxy serverStoreProxy = service.getServerStoreProxy(spaceIdentifier, new StoreConfigurationImpl<>(config, 1, null, null), Consistency.STRONG);
 
-    ClusteredTierClientEntity clientEntity = getEntity(serverStoreProxy);
+    SimpleClusteredTierClientEntity clientEntity = getEntity(serverStoreProxy);
 
     ClusteredStateRepository stateRepository = new ClusteredStateRepository(spaceIdentifier, "test", clientEntity);
 
@@ -124,10 +118,10 @@ public class ClusteredStateRepositoryReplicationTest {
     service.stop();
   }
 
-  private static ClusteredTierClientEntity getEntity(ServerStoreProxy clusteringService) throws NoSuchFieldException, IllegalAccessException {
+  private static SimpleClusteredTierClientEntity getEntity(ServerStoreProxy clusteringService) throws NoSuchFieldException, IllegalAccessException {
     Field entity = clusteringService.getClass().getDeclaredField("entity");
     entity.setAccessible(true);
-    return (ClusteredTierClientEntity)entity.get(clusteringService);
+    return (SimpleClusteredTierClientEntity)entity.get(clusteringService);
   }
 
 }
